Adapted the elements' update call in arena, adapted poweruptest with dt parameter in update, added dt parameter in totestelement's update

+++ src/main/java/it/unibo/pyxis/model/arena/Arena.java
@@ -17,7 +17,7 @@ public interface Arena {
      * @param delta
      *              The passed time.
      */
-    void update(Double delta);
+    void update(int delta);
 
     /**
      * Handle a {@link BrickDestructionEvent}.
